client_cam/requirements.txt
grpcio
grpcio-tools
opencv-python
